Fix GetLength() method to return an 'int', not a 'double'. Fix Rent method to remove '(int)' at return

+++ VHSTape.cs
@@ -40,7 +40,7 @@ namespace Activity4._4._3
         {
             return this._name;
         }
-        public double GetLength()
+        public int GetLength()
         {
             return this._length;
         }
